Feat: 개인정보, 마케팅 컨펌 뷰 컴포넌트 추가
